Update telegram bot service. Added message handler class to improve code readability and eliminate repetition. Also update constructors for TelegramBotPolling and TelegramBotWebhook classes.
